refactor: split icons into dedicated files
