+++ src/components/icons/MinusIcon.tsx
@@ -1,0 +1,3 @@
+import { FiMinus as MinusIcon } from 'react-icons/fi';
+
+export default MinusIcon;
