+++ src/components/icons/PlusIcon.tsx
@@ -1,0 +1,3 @@
+import { FiPlus as PlusIcon } from 'react-icons/fi';
+
+export default PlusIcon;
