+++ src/components/icons/ZapIcon.tsx
@@ -1,0 +1,3 @@
+import { FiZap as ZapIcon } from 'react-icons/fi';
+
+export default ZapIcon;
