--- src/components/icons/index.tsx
@@ -1,1 +1,0 @@
-export { FiPlus as PlusIcon, FiMinus as MinusIcon, FiZap as ZapIcon } from 'react-icons/fi';
